Task-44453: Make social share activity drawer common and reusable
