+++ webapp/portlet/src/main/webapp/share-activity-drawer/components/ShareActivitySpaceItem.vue
