+++ webapp/portlet/src/main/webapp/share-activity-drawer/initComponents.js
@@ -1,0 +1,14 @@
+import ShareActivityDrawer from './components/ShareActivityDrawer.vue';
+import ShareActivitySuggester from './components/ShareActivitySuggester.vue';
+import ShareActivitySpaceItem from './components/ShareActivitySpaceItem.vue';
+import ShareActivityNotificationAlerts from './components/ShareActivityNotificationAlerts.vue';
+
+const components = {
+  'share-activity-drawer': ShareActivityDrawer,
+  'share-activity-suggester': ShareActivitySuggester,
+  'share-activity-space-item': ShareActivitySpaceItem,
+  'share-activity-notification-alerts': ShareActivityNotificationAlerts,
+};
+for (const key in components) {
+  Vue.component(key, components[key]);
+}
